refactor: Make rest of the types match convention
