+++ src/common/entities/EWishlist.ts
@@ -1,0 +1,23 @@
+import { Prisma } from '@prisma/client';
+import { ELecture } from './ELecture';
+
+export namespace EWishlist {
+  export const WithLectures = Prisma.validator<Prisma.timetable_wishlistArgs>()(
+    {
+      include: {
+        timetable_wishlist_lectures: {
+          include: {
+            subject_lecture: {
+              include: ELecture.Details.include,
+            },
+          },
+          where: { subject_lecture: { deleted: false } },
+        },
+      },
+    },
+  );
+
+  export type WithLectures = Prisma.timetable_wishlistGetPayload<
+    typeof WithLectures
+  >;
+}
