+++ src/common/interfaces/serializer/wishlist.serializer.ts
@@ -1,9 +1,9 @@
-import { WishlistWithLectures } from '../../schemaTypes/types';
+import { EWishlist } from 'src/common/entities/EWishlist';
 import { WishlistWithLecturesResponseDto } from '../dto/wishlist/wishlist.response.dto';
 import { toJsonLecture } from './lecture.serializer';
 
 export const toJsonWishlist = (
-  wishlist: WishlistWithLectures,
+  wishlist: EWishlist.WithLectures,
 ): WishlistWithLecturesResponseDto => {
   return {
     lectures: wishlist.timetable_wishlist_lectures.map((lecture) =>
